1.8.9 Compatibility Fix
These changes might fix the issue that a lot of mods are having, where they inherit acceptedMinecraftVersion instead of setting it manually, and something goes wrong. Minecraft Errors and says that <mod> wants Minecraft [1.8,1.8]

As far as I can tell, the round bracket at the end is important, but I might not be.

I shall be testing this.